refactor: isPlayingState 를 isPlaying 으로 네미밍 수정
