+++ src/main/kotlin/blackjack/domain/Players.kt
@@ -2,5 +2,5 @@ package blackjack.domain
 
 class Players(private val players: Set<Player>) : Set<Player> by players {
     val countOfPlayingState: Int
-        get() = players.count { it.isPlayingState }
+        get() = players.count { it.isPlaying }
 }
